requirements.txt
openai>=1.0.0
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
moviepy>=1.0.3
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
pillow>=10.0.0
